Summarize data for player careers

This function is essentially a convenient wrapper for dplyr::summarise which provides a dataframe with the following columns: NumPunts, pEPA, Gross, Net, RERUN, SHARP_RERUN_OF, SHARP_RERUN_PD, first_year, last_year, team_logo_espn, team_color, team_color2. Additional columns may be added via valid calls to dplyr::summarise as additional arguments. For comparison of punter seasons, see puntr::by_punter_seasons; for comparison of punter games, see puntr::by_punter_games.

by_punters(punts, ..., threshold = 64)

Arguments

punts

The play-by-play punting data to be summarized

...

Any additional arguments will be passed through to dplyr::summarise

threshold

The minimum number of career punts needed to be included, defaults to 64

Value

A tibble where each row is a punter and each column is a stat

Examples

if (FALSE) {
punters <- by_punters(punts)
punters_custom <- by_punters(punts, longest_punt = max(GrossYards))
more_punters <- by_punters(punts, threshold=16)
}